converted notifications container to functional component
